refactor(agenda): move calendar navigation into its own component
