+++ components/Agenda/index.vue
@@ -2,8 +2,20 @@
   <div class="relative -top-24 z-10">
     <BaseContainer>
       <div class="bg-white p-10 rounded-xl shadow flex flex-col gap-7">
-        <AgendaCalendar :selected-date="selectedDate" @change="setSelectedDate" />
-        <AgendaList :selected-date="selectedDate" @change="setSelectedDate" />
+        <AgendaNavigation
+          :calendar-api="calendarApi"
+          :selected-date="selectedDate"
+          @change="setSelectedDate"
+        />
+        <AgendaCalendar
+          :calendar-api.sync="calendarApi"
+          :selected-date="selectedDate"
+          @change="setSelectedDate"
+        />
+        <AgendaList
+          :selected-date="selectedDate"
+          @change="setSelectedDate"
+        />
       </div>
     </BaseContainer>
   </div>
@@ -13,6 +25,7 @@
 export default {
   data () {
     return {
+      calendarApi: null,
       selectedDate: new Date()
     }
   },
